server/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-google-genai
